int64 indexing: complete migration from NumSharp
Cherry-picked changes from ilkernel branch (49254ad6) and completed the int64 migration:

Core Changes:
- Shape.GetOffset and GetSubshape now use long* for indices
- ArraySlice.Allocate has long overloads for count parameter
- UnmanagedStorage methods use long for size/offset/count
- Shape.Vector accepts long parameter

Removed int downcasts (previously using (int)size, (int)Count, etc.):
- UnmanagedStorage.cs: Allocate methods now use long
- UnmanagedStorage.Setters.cs: ReplaceData uses Shape.Vector(long)
- UnmanagedStorage.Cloning.cs: CloneData uses long count
- NDArray.cs: Shape.Vector(values.Count) without cast
- np.ones.cs, np.full.cs, np.full_like.cs: Allocate with shape.size
- np.meshgrid.cs: reshape with x.size without cast

Added overflow checks where .NET APIs require int:
- Slice.ToSliceDef: checks dim > int.MaxValue (SliceDef uses int)
- np.searchsorted: returns long indices, checks array size
- ILKernelGenerator.cs: Added DecimalZero/One/MinValue/MaxValue fields

Backward compatibility:
- int[] overloads delegate to long[] overloads (reshape, GetOffset)
- Transpose keeps int[] for axis permutation (max 32 dims)
- Explicit (int)Shape cast operator with overflow check

Test updates:
- Shape.Unmanaged.Tests.cs: Uses long* for GetOffset tests

+++ src/NumSharp.Core/Backends/Default/Indexing/Default.NonZero.cs
@@ -163,7 +163,7 @@ namespace NumSharp.Backends
 
             if (keepdims)
             {
-                var ks = new int[nd.ndim];
+                var ks = new long[nd.ndim];
                 for (int d = 0, sd = 0; d < nd.ndim; d++)
                     ks[d] = (d == axis) ? 1 : (sd < outputDims.Length ? outputDims[sd++] : 1);
                 result.Storage.Reshape(new Shape(ks));
